chore: fix issues with data persisting
